fix bug in options and don't set unused opts
Currently we pass a large number of options to window.open() but upon
investigation it seems a large majority of these options (all except
left, top, width, height) seem to do nothing on modern browsers (Edge,
Firefox, Chrome). In fact setting the 'location' option actually causes
Chrome to open the popup in a new tab rather than a new window.

Rather than risk similar bugs in the future I've just removed all of the
options that do not have an effect as they are essentially useless.