feat(redirect): modify callback.js to use postMessage
BREAKING CHANGE: callback.js now uses the `postMessage` Web API to
interact with the parent window instead of using `window.opener`. This
makes it much easier to use the SDK with single page applications where
it is common for the application and the back-end to live on different
origins. To upgrade all that needs to be done is to switch from
callback.js to redirect.js.

As a note this breaks compatibility for all browsers that do not support
`postMessage`. See https://caniuse.com/#feat=x-doc-messaging for more
details on supported browsers.