Fix jsdom localStorage bug
In order to test `postMessage` the window URL must be exposed to
`reconfigure` in order to post to arbitrary URIs. As Jest does not
normally allow window URL manipulation two additional packages are
needed - `jest-environment-jsdom` & `jest-environment-jsdom-global`.

This leads to an issue on Travis where we are accessing the browser's
`localStorage` which Travis does not have. This happens because the
default `testURL` value for jsdom is `about:blank`. To prevent this
issue overwrite the default `testURL` with localhost.